record WIP, start moving re-usable logic to helpers
The current users/get handler should be the body of "fetchUser" somewhere.
The user should get to choose whether to use or override this logic.

Currently the handler needs get pushed "up" into the processing
pipeline in order to become agnostic of many details implementation.
Ideally, anywhere in the server, one could ask for profile information
using the following:

server.fetchUser(name, myCallback);

This records moving some of the helper logic into places where it can
be more easily re-used, within the scope of a server, but not
necessarily in the context of a specific http request.